explicit error on unsupported query arg
