+++ query_parameters.go
@@ -1,0 +1,60 @@
+// Licensed to ClickHouse, Inc. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. ClickHouse, Inc. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package clickhouse
+
+import (
+	"github.com/ClickHouse/clickhouse-go/v2/lib/driver"
+	"github.com/pkg/errors"
+	"regexp"
+	"time"
+)
+
+var (
+	ErrExpectedStringValueInNamedValueForQueryParameter = errors.New("expected string value in NamedValue for query parameter")
+
+	hasQueryParamsRe = regexp.MustCompile("{.+:.+}")
+)
+
+func bindQueryOrAppendParameters(paramsProtocolSupport bool, options *QueryOptions, query string, timezone *time.Location, args ...interface{}) (string, error) {
+	// prefer native query parameters over legacy bind if query parameters provided explicit
+	if len(options.parameters) > 0 {
+		return query, nil
+	}
+
+	// validate if query contains a {<name>:<data type>} syntax, so it's intentional use of query parameters
+	// parameter values will be loaded from `args ...interface{}` for compatibility
+	if paramsProtocolSupport &&
+		len(args) > 0 &&
+		hasQueryParamsRe.MatchString(query) {
+		options.parameters = make(Parameters, len(args))
+		for _, a := range args {
+			if p, ok := a.(driver.NamedValue); ok {
+				if str, ok := p.Value.(string); ok {
+					options.parameters[p.Name] = str
+					continue
+				}
+			}
+
+			return "", ErrExpectedStringValueInNamedValueForQueryParameter
+		}
+
+		return query, nil
+	}
+
+	return bind(timezone, query, args...)
+}
